server/benefit: move benefit grant repository
